sony: Add Sony REST client
Add a Sony RESTClient for talking to Sony Bravia TVs using its REST IP
control protocol. This is a dodgy HTTP protocol (not really REST at all)
with poor security, but is a "modern" way to control a Sony Bravia TV.

There is also a line-based text control protocol from the olden days of
serial ports. This is not implemented, but may be in future if it turns
out to be faster.

Arguable this code could live in a package as it may be usable to other
projects. That will be considered later.

Subsequent commits will add methods to the RESTClient to perform
specific queries and controls on the TV as required by offscreen.

Link: https://pro-bravia.sony.net/develop/integrate/rest-api/spec/index.html