sony: Add `tv input` command to query/control inputs
Add a `tv input` command to query the currently selected input, query
all the available inputs and to change the selected input.

Inputs can be labelled on the TV set, in which case that label can be
used to select the input. Alternatively the non-human-friendly URI can
be used to select an input, which must be used for inputs without
labels.

e.g.

    $ offscreen tv input --list
    URI                        LABEL
    extInput:component?port=1
    extInput:composite?port=1
    extInput:hdmi?port=1       TVBox
    extInput:hdmi?port=2       edoras
    extInput:hdmi?port=3       palantr
    extInput:hdmi?port=4       orthanc
    extInput:widi?port=1
    $ offscreen tv input
    edoras
    $ offscreen tv input palantr
    # <selects input labelled with "palantr">
    $ offscreen tv power off
    $ offscreen tv input
    offscreen: error: selected input: decode: Display Is Turned off