sony: Add `tv toggle` command
Add the `tv toggle` command to toggle the state of the TV with a
heuristic based on the current state of the TV and the host from which
the command is being issued.

If the TV is off, toggle will turn on the set and select the input
labelled with the hostname.

If the TV is on and the currently selected input matches the hostname,
the TV set is turned off.

If the TV is on and the currently selected input does not match the
hostname, the input that is labelled with the hostname is selected.

As input labels are limited to seven characters, if the hostname is
longer than that, it will be truncated by taking the first six
characters and the last character. This seems to produce the most
recognisable labels. e.g. palantir -> palantr

This toggle command is intended to be bound to a hot key so when sitting
down in front of a computer, it can be pressed to say "make me active
now", but if already active, it is just turned off.