lint: Clean up lint settings
Disabled the `nonamedreturns` linter, because I believe that it is just
wrong. Named returns are ok in some cases and a computer will never be
able to tell which. So just turn it off and let human judgement rule.

Put `//nolint:goerr113` at the top of all source files that trigger
this. I don't want to turn this off altogether, but I think the problems
it reports are no problems when done in the `main` package.

Fix some incorrect indentation.