screen: Add run command to manage TV
Add the `run` command as the default command to manage a Sony TV to turn
it on and off in sync with the screen saver. It will turn off the TV if
the screen saver activates and the TV's current input matches that of
the host, and only when the TV is connected. It will turn on the TV if
the screen saver deactivates, and select the input for the connected
machine if not selected.