cli: Rework `--input` handling
Move the computation of the default input label that is based on the
hostname to a kong `PostBuild` step rather than doing it in the
`tv toggle` command handler. This allows kong to show the default value
in the help output and will also leave the code in a common place when
further commands are added with the `--input` flag.