cli: Rework `--input` handling
Rework the handling of the `--input` flag used in a couple of places to
set a default dynamically based on the hostname and to factor our common
code that turns an input label into an input URI for use with the Sony
API.

The default handling makes it possible for kong to show the default
value when showing help. The refactoring makes the code a bit cleaner.