screen: Add `list` command
Add a `list` command to list the manufacturer ID and product code of all
connected displays that have this info available. This is to be able to
discover what these values should be if the defaults are incorrect (the
defaults are set for the TV model that offscreen was developed for.