♻️ makefile: Move help descriptions to prior line (#4)
Update the `help` target awk script to look for the target description
for help messages on the line before the target instead of on the same
line. This is a more normal style for documenting things in code and
makes the target line shorter which can matter on some of the long
lines.

This also collects multi-line messages and combines it into a single
message, although that is dubious since you would not want overly long
messages.

It now uses `column(1)` to format the columns rather than relying on a
fixed width of 25 for the first column.

Pull-Request: https://github.com/foxygoat/offscreen/pull/4