♻️ screen: Handle returned error in ScreenWatcher (#7)
Add an `error` return value to the `ScreenWatcher.SSchange()` method and
return that from the `Screen.Watch()` function if an error is returned.
This is instead of panicking in the watcher function and then catching
it in `RunCmd.Run()`. I assume I used the panic method as I thought the
`ScreenWatcher` had no business returning an error, but I've rethought
this and now I think it is simpler.

As I later want to add error logging, making the error paths cleaner
will help.

Pull-Request: https://github.com/foxygoat/offscreen/pull/7