💚 ci: Fix goreleaser config after upgrade (#8)
Fix the goreleaser config after commit c4468b9e upgraded it from 1.12.3
to 1.22.1. Version 1.19.0 removed the deprecated `archives.replacements`
field, which could be replaced with `name_template`. However, I have no
good reason for the replacements - that part of the config was just
cargo-culted I think - so just get rid of the field altogether. This
will change the naming of the release artifacts on GitHub.

Also use `--clean` instead of `--rm-dist` in the `Makefile` as the
latter is deprecated and will likely break in a future upgrade.

Pull-Request: https://github.com/foxygoat/offscreen/pull/8