✨ config: Allow build-time config of bravia config (#6)
Pass the `$OFFSCREEN_HOSTNAME` and `$OFFSCREEN_PSK` env vars to the
build so they can be built into the binary. This makes the generated
binaries self-contained for a particular deployment meaning these vars
do not need to be set at run time.

I manage these vars in a `.envrc` file (for `direnv`) and do not have
them in my `.bashrc`. This change builds the values into the binary when
built locally (not on CI though so not in any released binaries). I can
then copy this binary to all machines connected to the Bravia TV so it
is preconfigured to talk to it.

The build-time vars are only used if the values are not specified at
run-time - either via env vars or the CLI directly.

Add some tests for this as I got the code wrong the first time. This
introduces `github.com/matryer/is` as a testing dependency.

Pull-Request: https://github.com/foxygoat/offscreen/pull/6